refactor: Extract progress display into MigrationProgressInfo component
Extract progress bar, remaining time, and cancel button from
NextcloudProgressDialog into a reusable component to improve code
organization and separation of concerns.
